requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
fire==0.5.0
fonttools==4.38.0
lxml==4.9.2
numpy==1.24.2
opencv-python==4.7.0.72
pdf2docx==0.5.6
PyMuPDF==1.21.1
python-decouple==3.8
python-docx==0.8.11
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
termcolor==2.2.0
